requirement.txt
Django==3.2
djangorestframework==3.10.3